# Use luatex in --luaonly mode to query kpsewhich.