--- a/doc/api/next_api_changes/deprecations/19558-AL.rst
+++ b/doc/api/next_api_changes/deprecations/19558-AL.rst
@@ -1,0 +1,3 @@
+The *format* parameter of ``dviread.find_tex_file``
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+... is deprecated (with no replacement).