--- a/lib/matplotlib/mpl-data/kpsewhich.lua
+++ b/lib/matplotlib/mpl-data/kpsewhich.lua
@@ -1,0 +1,3 @@
+-- see dviread._LuatexKpsewhich
+kpse.set_program_name("tex")
+while true do print(kpse.lookup(io.read():gsub("\r", ""))); io.flush(); end