@anntzer 2021-05-01
> this added dependency

Right now it's not a new dependency (we just silently fall back to the old slow path if luatex is not present); the question is indeed whether we want to emit a noisier warning or even just fail in that case and make luatex a hard dependency.  (I think there's some appeal to keep dependencies minimal?)
Also note that https://github.com/matplotlib/matplotlib/pull/19531 achieves the same result *without* adding a dependency on luatex, but doesn't work on Windows due to obscure buffering issues...