@anntzer 2021-05-21
I decided to just mention it in the docs and live with it, it's optional anyways.