@jklymak 2021-05-21
```
WARNING: error while formatting signature for matplotlib.texmanager.TexManager.get_rgba: Handler <function mangle_signature at 0x7fd341a9b200> for event 'autodoc-process-signature' threw an exception (exception: Unknown section Example in the docstring of <function TexManager.get_rgba at 0x7fd33a8397a0> in /home/circleci/project/lib/matplotlib/texmanager.py.)
```